feat(key-management-service): remove deprecated methods and packages ref: MANAGER-13837
