+++ packages/manager/apps/key-management-service/package.json
@@ -24,15 +24,13 @@
     "@ovh-ux/manager-config": "^7.3.0",
     "@ovh-ux/manager-core-api": "*",
     "@ovh-ux/manager-core-utils": "*",
-    "@ovh-ux/manager-react-core-application": "*",
     "@ovh-ux/manager-react-shell-client": "^0.4.0",
     "@ovh-ux/manager-tailwind-config": "*",
-    "@ovh-ux/shell": "^3.1.0",
     "@ovhcloud/manager-components": "*",
-    "@ovhcloud/ods-common-core": "17.1.0",
-    "@ovhcloud/ods-common-theming": "^17.1.0",
-    "@ovhcloud/ods-components": "^17.1.0",
-    "@ovhcloud/ods-theme-blue-jeans": "^17.1.0",
+    "@ovhcloud/ods-common-core": "17.2.1",
+    "@ovhcloud/ods-common-theming": "17.2.1",
+    "@ovhcloud/ods-components": "17.2.1",
+    "@ovhcloud/ods-theme-blue-jeans": "17.2.1",
     "@tanstack/react-query": "^5.12.2",
     "@tanstack/react-table": "^8.11.6",
     "i18next": "^20.4.0",
