+++ packages/manager/apps/key-management-service/src/App.tsx
@@ -1,10 +1,9 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useContext } from 'react';
 import { createHashRouter, RouterProvider } from 'react-router-dom';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { odsSetup } from '@ovhcloud/ods-common-core';
-import { useShell } from '@ovh-ux/manager-react-shell-client';
-import '@ovhcloud/ods-theme-blue-jeans';
+import { ShellContext } from '@ovh-ux/manager-react-shell-client';
 import queryClient from './query.client';
 
 import appRoutes from './routes';
@@ -13,9 +12,9 @@ odsSetup();
 const router = createHashRouter(appRoutes);
 
 function App() {
-  const shell = useShell();
+  const shellContext = useContext(ShellContext);
   useEffect(() => {
-    shell.ux.hidePreloader();
+    shellContext.shell.ux.hidePreloader();
   }, []);
 
   return (
