--- packages/manager/apps/key-management-service/src/index.scss
@@ -1,11 +1,0 @@
-.border-ods-primary-100 {
-  border-color: var(--ods-color-primary-100);
-}
-
-.background-ods-primary-100 {
-  background-color: var(--ods-color-primary-100);
-}
-
-.background-ods-primary-000 {
-  background-color: var(--ods-color-primary-000);
-}
