+++ packages/manager/apps/key-management-service/src/index.tsx
@@ -3,16 +3,14 @@ import ReactDOM from 'react-dom/client';
 import {
   ShellProvider,
   initShellContext,
+  initI18n,
 } from '@ovh-ux/manager-react-shell-client';
 import App from './App';
-import initI18n from './i18n';
 
 import './global.css';
+import '@ovhcloud/ods-theme-blue-jeans';
 
-const init = async (
-  appName: string,
-  { reloadOnLocaleChange } = { reloadOnLocaleChange: false },
-) => {
+const init = async (appName: string) => {
   const context = await initShellContext(appName);
 
   const region = context.environment.getRegion();
@@ -22,19 +20,11 @@ const init = async (
     // nothing to do
   }
 
-  const locales = await context.shell.i18n.getAvailableLocales();
-
-  const i18n = initI18n(
-    context.environment.getUserLocale(),
-    locales.map(({ key }: any) => key),
-  );
-
-  context.shell.i18n.onLocaleChange(({ locale }: { locale: string }) => {
-    if (reloadOnLocaleChange) {
-      window.top?.location.reload();
-    } else {
-      i18n.changeLanguage(locale);
-    }
+  await initI18n({
+    context,
+    reloadOnLocaleChange: true,
+    defaultNS: appName,
+    ns: [appName, `${appName}/listing`],
   });
 
   ReactDOM.createRoot(document.getElementById('root')!).render(
@@ -46,4 +36,4 @@ const init = async (
   );
 };
 
-init('key-management-service', { reloadOnLocaleChange: true });
+init('key-management-service');
