+++ packages/manager/apps/key-management-service/src/pages/dashboard/index.tsx
@@ -1,6 +1,29 @@
-import React from 'react';
-import Tabs1 from './Tabs1';
+import React, { Suspense } from 'react';
+import { useTranslation } from 'react-i18next';
+import { useResolvedPath } from 'react-router-dom';
+import Dashboard from '@/components/layout-helpers/Dashboard/Dashboard';
+import Loading from '@/components/Loading/Loading';
 
-export default function IndexTabs1() {
-  return <Tabs1 />;
+export default function DashboardPage() {
+  const { t } = useTranslation('key-management-service/dashboard');
+  const tabsList = [
+    {
+      name: 'general_infos',
+      title: t('general_informations'),
+      to: useResolvedPath('').pathname,
+    },
+    {
+      name: 'custom tab',
+      title: 'custom tab',
+      to: useResolvedPath('Tabs2').pathname,
+    },
+  ];
+
+  return (
+    <div>
+      <Suspense fallback={<Loading />}>
+        <Dashboard tabs={tabsList} />
+      </Suspense>
+    </div>
+  );
 }
