--- packages/manager/apps/key-management-service/src/shell.ts
@@ -1,18 +1,0 @@
-let shellClient: any;
-
-export const setShellClient = (client: any) => {
-  shellClient = client;
-
-  // set callbacks on locale change
-  shellClient.i18n.onLocaleChange(() => {
-    window.top?.location.reload();
-  });
-
-  return shellClient;
-};
-
-export const getShellClient = () => {
-  return shellClient;
-};
-
-export default { setShellClient, getShellClient };
